# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Personal AI Development Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /checkpoints:
    get:
      operationId: listCheckpoints
      tags: [checkpoints]
      summary: List checkpoints for a project
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Checkpoints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Checkpoint"

    post:
      operationId: createCheckpoint
      tags: [checkpoints]
      summary: Create a checkpoint (git commit)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckpointInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Checkpoint"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /checkpoints/{id}/rollback:
    post:
      operationId: rollbackCheckpoint
      tags: [checkpoints]
      summary: Rollback to a checkpoint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Rolled back
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── BUILD / PREVIEW ───────────────────────────────────────────────────────
  /build/run:
    post:
      operationId: runBuild
      tags: [build]
      summary: Start a build for a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GitProjectRef"
      responses:
        "200":
          description: Build started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuildStatus"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /build/status:
    get:
      operationId: getBuildStatus
      tags: [build]
      summary: Get build status for a project
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Build status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuildStatus"
  /preview/start:
    post:
      operationId: startPreview
      tags: [build]
      summary: Start the preview server for a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GitProjectRef"
      responses:
        "200":
          description: Preview started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreviewStatus"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /preview/stop:
    post:
      operationId: stopPreview
      tags: [build]
      summary: Stop the preview server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GitProjectRef"
      responses:
        "200":
          description: Stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreviewStatus"
  /preview/status:
    get:
      operationId: getPreviewStatus
      tags: [build]
      summary: Get preview server status
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Preview status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreviewStatus"

  # ── AI ────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean

    # Auth
...
    GitProjectRef:
      type: object
      required: [projectId]
      properties:
        projectId:
          type: number
...
    Checkpoint:
      type: object
      required: [id, projectId, label, commitHash, createdAt]
      properties:
        id:
          type: number
        projectId:
          type: number
        label:
          type: string
        summary:
          type: ["string", "null"]
        commitHash:
          type: string
        filesChanged:
          type: ["integer", "null"]
        buildPassed:
          type: ["boolean", "null"]
        createdAt:
          type: string
    CheckpointInput:
      type: object
      required: [projectId, label]
      properties:
        projectId:
          type: number
        label:
          type: string
        summary:
          type: string

    # Build / Preview
    BuildStatus:
      type: object
      required: [projectId, state]
      properties:
        projectId:
          type: number
        state:
          type: string
          enum: [idle, running, success, failed]
        exitCode:
          type: ["integer", "null"]
        startedAt:
          type: ["string", "null"]
        finishedAt:
          type: ["string", "null"]
    PreviewStatus:
      type: object
      required: [projectId, state]
      properties:
        projectId:
          type: number
        state:
          type: string
          enum: [stopped, starting, running, crashed]
        port:
          type: ["integer", "null"]
        pid:
          type: ["integer", "null"]
        url:
          type: ["string", "null"]

    # AI
...