# lib/api-zod/src/generated/types/buildStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Personal AI Development Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { BuildStatusState } from './buildStatusState';

export interface BuildStatus {
  projectId: number;
  state: BuildStatusState;
  /** @nullable */
  exitCode?: number | null;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  finishedAt?: string | null;
}